Move stuff to core/control.cpp
